pyproject.toml
[project]
name = "nl-sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
    "sqlalchemy>=2.0.44",
    "sqlglot>=28.1.0",
    "streamlit>=1.52.0",
]
